% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdQmat.R
\name{birth_death_Q}
\alias{birth_death_Q}
\title{Populate a transition rate matrix for a Birth-Death process given birth and death rates}
\usage{
birth_death_Q(lambda, mu, m)
}
\arguments{
\item{lambda}{birth rate}

\item{mu}{death rate}

\item{m}{maximum number of population i.e. carrying capacity or in other contexts maximum number of states}
}
\value{
matrix
}
\description{
{For convenience to populate transition rate matrices for common CTMC models.}
}
