% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitting_times_dtmc.R
\name{ctmc_hitting_times}
\alias{ctmc_hitting_times}
\title{Compute expected hitting times until reaching a state}
\usage{
ctmc_hitting_times(Q, states)
}
\arguments{
\item{Q}{original probability transition matrix}

\item{states}{the set of states to hit}
}
\value{
vector/matrix
}
\description{
{Solve the linear system resulting from
first step analysis until time of absorption.}
}
