% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtmc_stationary.R
\name{ctmc_steady_state}
\alias{ctmc_steady_state}
\title{Compute the stationary distribution of a CTMC given a transition-rate matrix}
\usage{
ctmc_steady_state(Q)
}
\arguments{
\item{Q}{the transition-rate matrix for the given CTMC}
}
\value{
vector
}
\description{
{Uses the basic linear-algebra constrained method to solve for the limiting distribution, if it exists.}
}
