% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitting_times_dtmc.R
\name{dtmc_hitting_chances}
\alias{dtmc_hitting_chances}
\title{Compute expected hitting chance of reaching a state}
\usage{
dtmc_hitting_chances(P, states, before = FALSE)
}
\arguments{
\item{P}{original probability transition matrix}

\item{states}{the set of states to hit}

\item{before}{boolean to compute hitting probability of state i before j}
}
\value{
vector/matrix
}
\description{
{Solve the linear system resulting from
first step analysis until time of absorption.}
}
