% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitting_times_dtmc.R
\name{dtmc_hitting_times}
\alias{dtmc_hitting_times}
\title{Compute expected hitting times until reaching a state}
\usage{
dtmc_hitting_times(P, states)
}
\arguments{
\item{P}{original probability transition matrix}

\item{states}{the set of states to hit}
}
\value{
vector/matrix
}
\description{
{Solve the linear system resulting from
first step analysis until time of absorption.}
}
