% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtmc_stationary.R
\name{dtmc_stationary}
\alias{dtmc_stationary}
\title{Compute the stationary distribution of a DTMC given a transition-probability matrix}
\usage{
dtmc_stationary(P)
}
\arguments{
\item{P}{the transition-probability matrix for the given DTMC}
}
\value{
vector
}
\description{
{Uses the basic linear-algebra constrained method to solve for the limiting distribution, if it exists.}
}
