% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ctmc.R
\name{estimate_ctmc}
\alias{estimate_ctmc}
\title{Estimate the transition-rate matrix of an observed CTMC}
\usage{
estimate_ctmc(chain, times, m)
}
\arguments{
\item{chain}{the observed chain}

\item{times}{the times the chain is observed at}

\item{m}{the maximum of number of states in the chain}
}
\value{
matrix
}
\description{
{The scheme depends on number of transitions form a fixed state to any other state... more details to come.}
}
