% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dtmc.R
\name{estimate_dtmc}
\alias{estimate_dtmc}
\title{Estimate the transition-probability matrix of an observed DTMC}
\usage{
estimate_dtmc(chain, m)
}
\arguments{
\item{chain}{the observed chain}

\item{m}{the maximum of number of states in the chain}
}
\value{
matrix
}
\description{
{The scheme depends on number of transitions from a fixed state to any other state... more details to come.}
}
