% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamOnRye-data.R
\docType{data}
\name{hamOnRye}
\alias{hamOnRye}
\title{Ham on Rye}
\format{
a data.frame containing massive text (characters/strings)
}
\usage{
data(hamOnRye)
}
\description{
The entire Ham on Rye text
}
\keyword{datasets}
