% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_state.R
\name{initial_state}
\alias{initial_state}
\title{Simulate the initial state given an initial measure/distribution}
\usage{
initial_state(mu)
}
\arguments{
\item{mu}{the initial probability measure, a vector whose entries sum to unity}
}
\value{
integer which represents the state
}
\description{
{A basic sampling scheme based of uniform RVs and the given measure.}
}
