% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_chain.R
\name{jump_chain}
\alias{jump_chain}
\title{Convert between transition-rate matrix to the associated jump-chain's transition-probability matrix for a CTCM}
\usage{
jump_chain(Q)
}
\arguments{
\item{Q}{the transition-rate matrix given for the CTCM}
}
\value{
matrix
}
\description{
{Every CTMC has an embedded DTMC associated with it called the jump-chain. This function computes that jump
-chain's transition-probability matrix.}
}
