% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_sonnet.R
\name{markov_sonnet}
\alias{markov_sonnet}
\title{Generate a first-order Markov chain Shakespearean sonnet}
\usage{
markov_sonnet(sonnet_data, P = NULL)
}
\arguments{
\item{sonnet_data}{the object returned from \code{observe_sonnet_chain()}, a list containing \code{wordChain}, \code{encodedChain}, and \code{stateSpace}}

\item{P}{the probability transition matrix, if not passed will be estimated from the \code{sonnet_data} dataset}
}
\value{
string/character
}
\description{
{The transition probability matrix is estimated from all 154 sonnets}
}
