% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_text.R
\name{markov_text}
\alias{markov_text}
\title{Generate a first-order Markov chain Shakespearean sonnet}
\usage{
markov_text(text_data, P = NULL, wordsPerLine = 10, numLines = 14)
}
\arguments{
\item{text_data}{the object returned from \code{observe_text()}, a list containing \code{wordChain}, \code{encodedChain}, and \code{stateSpace}}

\item{P}{the probability transition matrix, if not passed will be estimated from the \code{sonnet_data} dataset}

\item{wordsPerLine}{the number of words per line}

\item{numLines}{the number of lines}
}
\value{
string/character
}
\description{
{The transition probability matrix is estimated from all 154 sonnets}
}
