% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_state.R
\name{next_state}
\alias{next_state}
\title{Simulate the next state given the current state of a DTMC}
\usage{
next_state(i, P)
}
\arguments{
\item{i}{the current state}

\item{P}{the transition-probability matrix}
}
\value{
integer
}
\description{
{Basic intervallic partition sampling of the measures.}
}
