% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_ham_on_rye.R
\name{observe_ham_on_rye}
\alias{observe_ham_on_rye}
\title{Convert the raw text of all of Ham on Rye to observed Markov chain of words}
\usage{
observe_ham_on_rye(punctuation = FALSE, uppercase = FALSE, reducedSize = 10000)
}
\arguments{
\item{punctuation}{boolean whether to keep punctuation (TRUE) or not (FALSE)}

\item{uppercase}{boolean whether to distinuqish between uppercase (TRUE) or not (FALSE)}

\item{reducedSize}{this text is huge, so using only portions of it makes the DTMC estimation go faster.}
}
\value{
list containing named entries
}
\description{
{Formatting of the raw text file of Ham on Rye. The function prepares it into a discrete time observed
chain by putting each word in a big vector/list, removing punctuation, cases, and titles.}
}
