% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_sonnet_chain.R
\name{observe_sonnet_chain}
\alias{observe_sonnet_chain}
\title{Convert the raw text of all 154 sonnets to an observed Markov chain of words}
\usage{
observe_sonnet_chain(punctuation = FALSE, uppercase = FALSE)
}
\arguments{
\item{punctuation}{boolean whether to keep punctuation (TRUE) or not (FALSE)}

\item{uppercase}{boolean whether to distinuqish between uppercase (TRUE) or not (FALSE)}
}
\value{
list containing named entries
}
\description{
{Formatting of the raw text file of 154 sonnets by Shakespeare. The function prepares it into a discrete time observed
chain by putting each word in a big vector/list, removing punctuation, cases, and titles.}
}
