% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctmc.R
\name{rctmc}
\alias{rctmc}
\title{Simulate a continuous time Markov chain}
\usage{
rctmc(tt, Q, mu, states = NULL)
}
\arguments{
\item{tt}{how long to simulate the chain for}

\item{Q}{the transition rate matrix also called the infinitesimal generator matrix}

\item{mu}{the initial distribution vector}

\item{states}{optional list of states for non-integer chains}
}
\value{
data.frame
}
\description{
{The standard scheme to simulate CTMCs, e.g. in J.R. Norris's textbook.}
}
