% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdtmc.R
\name{rdtmc}
\alias{rdtmc}
\title{Simulate a discrete time Markov chain}
\usage{
rdtmc(n, P, mu, states = NULL)
}
\arguments{
\item{n}{number of steps to simulate in the chain}

\item{P}{the probability transition matrix}

\item{mu}{the initial distribution vector}

\item{states}{optional states if simulating strings or non-integer chains}
}
\value{
data.frame
}
\description{
{The basic Markov chain simulation scheme found in J.R. Norris' book.}
}
