% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonnets-data.R
\docType{data}
\name{sonnets}
\alias{sonnets}
\title{Sonnets}
\format{a data.frame containing massive text (characters/strings)}
\source{
\href{https://www.opensourceshakespeare.org/views/sonnets/sonnet_view.php?Sonnet=all&pleasewait=1&msg=pl}{Open Source Shakespeare}
}
\usage{
data(sonnets)
}
\description{
All of Shakespeare's sonnets in one long text/data.frame, line by line
}
\keyword{datasets}
